+++ CHANGELOG.md
@@ -8,17 +8,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- `SnowflakeConnector` block - [#24](https://github.com/PrefectHQ/prefect-snowflake/pull/24)
 
 ### Changed
+- Moved the keywords, `database` and `warehouse`, from `SnowflakeCredentials` into `SnowflakeConnector` - [#24](https://github.com/PrefectHQ/prefect-snowflake/pull/24)
+- Moved the method `get_connection` from `SnowflakeCredentials` into `SnowflakeConnector` - [#24](https://github.com/PrefectHQ/prefect-snowflake/pull/24)
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
-- Swap out `block_initialization` for `_get_connect_params` so logging the block doesn't show secrets - [#24](https://github.com/PrefectHQ/prefect-snowflake/pull/24)
 
 ### Security
+- Fixed hiding the input password under `connect_params` when logging `SnowflakeCredentials` - [#24](https://github.com/PrefectHQ/prefect-snowflake/pull/24)
+
+## 0.2.0
+Released on August ??th, 2022.
 
 ## 0.1.3
 Released on July 26th, 2022.